Create favicon, logo.Rename pages
