+++ src/Components/Header/header.css
@@ -1,0 +1,19 @@
+.header {
+  display: flex;
+  justify-content: center;
+  margin-left: 50px;
+  margin-right: 50px;
+  padding-top: 20px;
+  gap: 30px;
+}
+  .logo {
+    background-image: url("/img/cesnek.png");
+    background-size: contain;
+    border-radius: 100%;
+    flex-basis: 50px;
+    height: 45px;
+    width: 45px;
+    position: absolute;
+    top: 10px;
+    left: 10px;
+}
