+++ src/Pages/HomePage/home.tsx
@@ -1,10 +1,11 @@
 import './home.css';
-import LinkButton from '../Components/LinkButton/linkButton';
+import LinkButton from '../../Components/LinkButton/linkButton';
 import React from 'react';
 
 export const HomePage: React.FC = () => (
   <div className="container">
-    <LinkButton to="/workers">Evidence brigádníků</LinkButton>
-    <LinkButton to="/work">Evidence práce</LinkButton>
+    <LinkButton to="/reports">Přehledy</LinkButton>
+    <LinkButton to="/work">Zadat práci</LinkButton>
   </div>
 );
+export default HomePage;
