+++ src/Pages/Reports/reports.tsx
@@ -1,0 +1,3 @@
+import './reports.css';
+
+export const Reports: React.FC = () => <></>;
