--- src/Pages/Workers/workers.tsx
@@ -1,3 +1,0 @@
-import './workers.css';
-
-export const Workers: React.FC = () => <></>;
