+++ src/index.css
@@ -24,14 +24,14 @@ a:hover {
 
 body {
   margin: 0;
-  display: flex;
+  width: 100%;
   place-items: center;
   min-width: 320px;
   min-height: 100vh;
 }
 
 h1 {
-  font-size: 3.2em;
+  font-size: 3em;
   line-height: 1.1;
 }
